refactor Generator into use case layer
