+++ src/main/java/entities/MazeFactory.java
@@ -6,7 +6,7 @@ package entities;
  */
 public class MazeFactory {
 
-    public MazeFacade create(int rows, int columns) {
-        return new MazeFacade(rows, columns);
+    public Maze create(int rows, int columns) {
+        return new Maze(rows, columns);
     }
 }
