+++ src/main/java/usecases/generateMazeUseCase/Generator.java
@@ -1,4 +1,6 @@
-package entities;
+package usecases.generateMazeUseCase;
+
+import entities.Maze;
 
 public interface Generator {
     public void generateMaze(Maze maze);
